requirements.txt
pyshark
pycryptodome